requirements.txt
eventlet==0.33.3
faiss-cpu==1.7.4
Flask==2.3.3
Flask-SocketIO==5.5.1
gunicorn==21.2.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7